CHE-3930: Move SshServiceClient from wsmaster to wsagent war
