+++ assembly/assembly-wsagent-war/src/main/java/com/codenvy/ext/java/server/MachineModule.java
@@ -22,7 +22,6 @@ import com.google.gson.JsonParser;
 import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
 import com.google.inject.assistedinject.FactoryModuleBuilder;
-import com.google.inject.multibindings.MapBinder;
 import com.google.inject.name.Names;
 
 import org.eclipse.che.EventBusURLProvider;
@@ -46,11 +45,8 @@ import org.eclipse.che.api.git.GitUserResolver;
 import org.eclipse.che.api.git.LocalGitUserResolver;
 import org.eclipse.che.api.project.server.ProjectApiModule;
 import org.eclipse.che.api.project.server.ProjectServiceLinksInjector;
-import org.eclipse.che.api.ssh.server.HttpSshServiceClient;
-import org.eclipse.che.api.ssh.server.SshServiceClient;
 import org.eclipse.che.api.user.server.spi.PreferenceDao;
 import org.eclipse.che.commons.lang.Pair;
-import org.eclipse.che.dto.server.DtoFactory;
 import org.eclipse.che.everrest.CheAsynchronousJobPool;
 import org.eclipse.che.git.impl.jgit.JGitConnectionFactory;
 import org.eclipse.che.ide.ext.microsoft.server.inject.MicrosoftModule;
@@ -58,6 +54,8 @@ import org.eclipse.che.inject.DynaModule;
 import org.eclipse.che.plugin.github.server.inject.GitHubModule;
 import org.eclipse.che.plugin.maven.generator.archetype.ArchetypeGenerator;
 import org.eclipse.che.plugin.maven.server.inject.MavenModule;
+import org.eclipse.che.plugin.ssh.key.HttpSshServiceClient;
+import org.eclipse.che.plugin.ssh.key.SshServiceClient;
 import org.eclipse.che.security.oauth.RemoteOAuthTokenProvider;
 import org.eclipse.che.security.oauth1.RemoteOAuthAuthorizationHeaderProvider;
 import org.everrest.core.impl.async.AsynchronousJobPool;
