+++ plugins/plugin-bitbucket/codenvy-plugin-bitbucket-ext-bitbucket/pom.xml
@@ -79,10 +79,6 @@
         </dependency>
         <dependency>
             <groupId>org.eclipse.che.core</groupId>
-            <artifactId>che-core-api-ssh</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.che.core</groupId>
             <artifactId>che-core-api-ssh-shared</artifactId>
         </dependency>
         <dependency>
